backend/requirements.txt
flask
flask-cors
werkzeug
mysql-connector-python>=8.0.31
paramiko>=3.0.0
pycryptodome>=3.16.0
cryptography>=38.0.3
marshmallow 
hypercorn>=0.14.3
paramiko-expect>=0.3.5
pynacl>=1.5.0
requests>=2.28.0
croniter>=1.3.8
pytest>=7.0
pytest-cov>=4.0
quart>=0.18.3
quart-cors>=0.6.0